requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.7

# Caching
redis==4.6.0
fastapi-cache2==0.2.1

# ML & Data
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0

# HTTP Client
requests==2.31.0
httpx==0.25.0

# Utilities
python-multipart==0.0.6
loguru==0.7.0
tqdm==4.66.1

Faker==19.6.2

# For CUDA 12.1 (NVIDIA GPUs)
torch==2.3.1 --extra-index-url https://download.pytorch.org/whl/cu121

# For CPU-only version
torch==2.3.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu